refactor: replace movie detail styling with new styled components and separate MovieContent component for improved organization and readability
