+++ src/pages/movie-detail/movie-detail.tsx
@@ -3,22 +3,8 @@ import { getMovie } from '@movie/services/api/movie-api'
 import { useService } from '@app/lib/use-service';
 import { ErrorComponent } from '@core/error/components/error-component';
 import { MovieDetailSkeleton } from '@app/lib/components/skeleton-movie-detail/skeleton-movie-detail';
-import { BackdropContainer, BackdropImage, Container, BackButton, MovieGrid, Poster, MovieInfo, Title, MovieDetails, DetailItem, ReleaseDate, Rating, Overview } from './movie-detail.stiled';
-
-
-const formatRuntime = (minutes: number): string => {
-  const hours = Math.floor(minutes / 60);
-  const remainingMinutes = minutes % 60;
-  return `${hours}h ${remainingMinutes}m`;
-};
-
-const formatPopularity = (popularity: number): string => {
-  return popularity.toFixed(1);
-};
-
-const formatLanguage = (language: string): string => {
-  return language.toUpperCase();
-};
+import { BackdropContainer, BackdropImage, Container, BackButton, MovieGrid } from './movie-detail.styled';
+import { MovieContent } from './components/movie-content/movie-content';
 
 export const MovieDetail: React.FC = () => {
   const { id } = useParams<{ id: string }>();
@@ -43,33 +29,7 @@ export const MovieDetail: React.FC = () => {
           ← Back to Movies
         </BackButton>
         <MovieGrid>
-          <Poster 
-            src={`https://image.tmdb.org/t/p/w500${movie.posterPath}`} 
-            alt={movie.title}
-          />
-          <MovieInfo>
-            <Title>{movie.title}</Title>
-            <MovieDetails>
-              {movie.language && (
-                <DetailItem>
-                  {formatLanguage(movie.language)}
-                </DetailItem>
-              )}
-              {movie.runtime && (
-                <DetailItem>
-                  {formatRuntime(movie.runtime)}
-                </DetailItem>
-              )}
-              <DetailItem>
-                Popularity: {formatPopularity(movie.popularity)}
-              </DetailItem>
-            </MovieDetails>
-            <ReleaseDate>Release Date: {movie.releaseDate}</ReleaseDate>
-            <Rating>
-              Rating: {movie.voteAverage.toFixed(1)}/10
-            </Rating>
-            <Overview>{movie.overview}</Overview>
-          </MovieInfo>
+          <MovieContent movie={movie} />
         </MovieGrid>
       </Container>
     </>
